Inject channel links in show notes
Each message in the current channel is parsed and only 'pure' link messages are kept. Note that using Message#getBlocks() seems safer to use, but Message#getText() is simpler for the time being.

Threaded messages (link replies) are not yet retrieved: they require an additional call to the Slack API.