Publish show notes to a GitHub repository
Unfortunately it is not possible to publish notes to a GitHub wiki through the GitHub API although wikis are git repositories (https://stackoverflow.com/questions/2402413/update-a-github-project-wiki-through-the-github-api). For the record, I tried to suffixing the repo name in the API URL with '.wiki' and '.wiki.git' but it did not work.